Remove evaluator code, update tests
