+++ timeseries/src/autogluon/timeseries/models/abstract/abstract_timeseries_model.py
@@ -103,6 +103,18 @@ class AbstractTimeSeriesModel(AbstractModel):
         self.freq: str = freq
         self.prediction_length: int = prediction_length
         self.quantile_levels = kwargs.get("quantile_levels", [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9])
+
+        if not all(0 < q < 1 for q in self.quantile_levels):
+            raise ValueError("Invalid quantile_levels specified. Quantiles must be between 0 and 1 (exclusive).")
+
+        # We ensure that P50 forecast is always among the "raw" predictions generated by _predict.
+        # We remove P50 from the final predictions if P50 wasn't present among the specified quantile_levels.
+        if 0.5 not in self.quantile_levels:
+            self.must_drop_median = True
+            self.quantile_levels = sorted(set([0.5] + self.quantile_levels))
+        else:
+            self.must_drop_median = False
+
         self._oof_predictions: Optional[List[TimeSeriesDataFrame]] = None
 
     def __repr__(self) -> str:
@@ -253,22 +265,6 @@ class AbstractTimeSeriesModel(AbstractModel):
                 "as hyperparameters when initializing or use `hyperparameter_tune` instead."
             )
 
-    def _check_predict_inputs(
-        self,
-        data: TimeSeriesDataFrame,
-        quantile_levels: Optional[List[float]] = None,
-        **kwargs,  # noqa: F841
-    ):
-        logger.debug(f"Predicting with time series model {self.name}")
-        logger.debug(
-            f"\tProvided data for prediction with {len(data)} rows, {data.num_items} items. "
-            f"Average time series length is {len(data) / data.num_items:.1f}."
-        )
-
-        quantiles = quantile_levels or self.quantile_levels
-        if not all(0 < q < 1 for q in quantiles):
-            raise ValueError("Invalid quantile value specified. Quantiles must be between 0 and 1 (exclusive).")
-
     def predict(
         self,
         data: Union[TimeSeriesDataFrame, Dict[str, TimeSeriesDataFrame]],
@@ -290,13 +286,6 @@ class AbstractTimeSeriesModel(AbstractModel):
         known_covariates : Optional[TimeSeriesDataFrame]
             A TimeSeriesDataFrame containing the values of the known covariates during the forecast horizon.
 
-        Other Parameters
-        ----------------
-        quantile_levels
-            Quantiles of probabilistic forecasts, if probabilistic forecasts are implemented by the
-            corresponding subclass. If None, `self.quantile_levels` will be used instead,
-            if provided during initialization.
-
         Returns
         -------
         predictions: TimeSeriesDataFrame
@@ -304,6 +293,21 @@ class AbstractTimeSeriesModel(AbstractModel):
             data is given as a separate forecast item in the dictionary, keyed by the `item_id`s
             of input items.
         """
+        predictions = self._predict(data=data, known_covariates=known_covariates, **kwargs)
+        if "0.5" in predictions.columns:
+            if self.eval_metric.optimized_by_median:
+                predictions["mean"] = predictions["0.5"]
+            if self.must_drop_median:
+                predictions = predictions.drop("0.5", axis=1)
+        return predictions
+
+    def _predict(
+        self,
+        data: Union[TimeSeriesDataFrame, Dict[str, TimeSeriesDataFrame]],
+        known_covariates: Optional[TimeSeriesDataFrame] = None,
+        **kwargs,
+    ) -> TimeSeriesDataFrame:
+        """Private method for `predict`. See `predict` for documentation of arguments."""
         raise NotImplementedError
 
     def _score_with_predictions(
@@ -320,7 +324,6 @@ class AbstractTimeSeriesModel(AbstractModel):
             prediction_length=self.prediction_length,
             target=self.target,
             seasonal_period=self.eval_metric_seasonal_period,
-            quantile_levels=self.quantile_levels,
         )
 
     def score(self, data: TimeSeriesDataFrame, metric: Optional[str] = None) -> float:
